arreglo de moneda

+++ tienda/models.py
@@ -11,7 +11,8 @@ class Producto(models.Model):
     
     nombre = models.CharField(max_length=100)
     descripcion = models.TextField()
-    precio = models.DecimalField(max_digits=10, decimal_places=2)
+    precio = models.IntegerField()
+
     imagen = models.ImageField(upload_to='productos/')
     categoria = models.CharField(max_length=20, choices=CATEGORIAS)
     stock = models.IntegerField(default=0)
